# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Faso Sènè Agribusiness Marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: prices
    description: Daily price updates
  - name: suppliers
    description: Supplier management
  - name: orders
    description: Customer orders
  - name: deliveries
    description: Delivery management
  - name: admin
    description: Admin dashboard

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # Products
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Daily Prices
  /prices:
    get:
      operationId: listPrices
      tags: [prices]
      summary: List current daily prices
      parameters:
        - name: productName
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of daily prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPrice"
    post:
      operationId: createPrice
      tags: [prices]
      summary: Create or update a daily price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyPriceInput"
      responses:
        "201":
          description: Created price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPrice"

  /prices/latest:
    get:
      operationId: getLatestPrices
      tags: [prices]
      summary: Get the latest prices for all products
      responses:
        "200":
          description: Latest prices per product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPrice"

  # Suppliers
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Register a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  /suppliers/{id}/products:
    get:
      operationId: getSupplierProducts
      tags: [suppliers]
      summary: Get products for a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Products for supplier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /suppliers/{id}/orders:
    get:
      operationId: getSupplierOrders
      tags: [suppliers]
      summary: Get orders for a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Orders for supplier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Deliveries
  /deliveries:
    get:
      operationId: listDeliveries
      tags: [deliveries]
      summary: List all deliveries
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deliveries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Delivery"
    post:
      operationId: createDelivery
      tags: [deliveries]
      summary: Create a delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryInput"
      responses:
        "201":
          description: Created delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delivery"

  /deliveries/{id}:
    patch:
      operationId: updateDelivery
      tags: [deliveries]
      summary: Update delivery status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryUpdate"
      responses:
        "200":
          description: Updated delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delivery"

  # Admin dashboard stats
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, category, unit, supplierId, isAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameLocal:
          type: ["string", "null"]
        category:
          type: string
        unit:
          type: string
        description:
          type: ["string", "null"]
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        currentPrice:
          type: ["number", "null"]
        isAvailable:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, category, unit, supplierId]
      properties:
        name:
          type: string
        nameLocal:
          type: string
        category:
          type: string
        unit:
          type: string
        description:
          type: string
        supplierId:
          type: integer
        isAvailable:
          type: boolean
        imageUrl:
          type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        nameLocal:
          type: string
        category:
          type: string
        unit:
          type: string
        description:
          type: string
        isAvailable:
          type: boolean
        imageUrl:
          type: string

    DailyPrice:
      type: object
      required: [id, productName, pricePerUnit, unit, date, updatedAt]
      properties:
        id:
          type: integer
        productName:
          type: string
        pricePerUnit:
          type: number
        unit:
          type: string
        date:
          type: string
        trend:
          type: ["string", "null"]
        percentChange:
          type: ["number", "null"]
        updatedAt:
          type: string

    DailyPriceInput:
      type: object
      required: [productName, pricePerUnit, unit, date]
      properties:
        productName:
          type: string
        pricePerUnit:
          type: number
        unit:
          type: string
        date:
          type: string
        trend:
          type: string
        percentChange:
          type: number

    Supplier:
      type: object
      required: [id, name, phone, region, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        region:
          type: string
        village:
          type: ["string", "null"]
        type:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        totalProducts:
          type: ["integer", "null"]
        totalOrders:
          type: ["integer", "null"]
        createdAt:
          type: string

    SupplierInput:
      type: object
      required: [name, phone, region, type]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        region:
          type: string
        village:
          type: string
        type:
          type: string
        description:
          type: string

    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        region:
          type: string
        village:
          type: string
        type:
          type: string
        description:
          type: string
        status:
          type: string

    Order:
      type: object
      required: [id, customerName, customerPhone, items, totalAmount, status, deliveryAddress, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        status:
          type: string
        deliveryAddress:
          type: string
        deliveryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        whatsappOrder:
          type: boolean
        createdAt:
          type: string

    OrderItem:
      type: object
      required: [productId, productName, quantity, unit, unitPrice, subtotal]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        subtotal:
          type: number

    OrderInput:
      type: object
      required: [customerName, customerPhone, items, deliveryAddress]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
        notes:
          type: string
        whatsappOrder:
          type: boolean

    OrderItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitPrice:
          type: number

    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        deliveryDate:
          type: string
        notes:
          type: string

    Delivery:
      type: object
      required: [id, orderId, driverName, driverPhone, status, scheduledDate, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        driverName:
          type: string
        driverPhone:
          type: string
        status:
          type: string
        scheduledDate:
          type: string
        deliveredAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    DeliveryInput:
      type: object
      required: [orderId, driverName, driverPhone, scheduledDate]
      properties:
        orderId:
          type: integer
        driverName:
          type: string
        driverPhone:
          type: string
        scheduledDate:
          type: string
        notes:
          type: string

    DeliveryUpdate:
      type: object
      properties:
        status:
          type: string
        deliveredAt:
          type: string
        notes:
          type: string

    AdminStats:
      type: object
      required: [totalSuppliers, totalProducts, totalOrders, totalRevenue, pendingOrders, activeDeliveries, todayOrders, topProducts]
      properties:
        totalSuppliers:
          type: integer
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        activeDeliveries:
          type: integer
        todayOrders:
          type: integer
        topProducts:
          type: array
          items:
            type: object
            required: [name, orderCount]
            properties:
              name:
                type: string
              orderCount:
                type: integer