# lib/api-zod/src/generated/types/dailyPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Faso Sènè Agribusiness Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyPrice {
  id: number;
  productName: string;
  pricePerUnit: number;
  unit: string;
  date: string;
  /** @nullable */
  trend?: string | null;
  /** @nullable */
  percentChange?: number | null;
  updatedAt: string;
}