# lib/api-zod/src/generated/types/dailyPriceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Faso Sènè Agribusiness Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyPriceInput {
  productName: string;
  pricePerUnit: number;
  unit: string;
  date: string;
  trend?: string;
  percentChange?: number;
}